+++ src/map/params.rs
@@ -1,2 +1,2 @@
-pub const CHUNK_LOG_SIZE: u64 = 3;
-pub const NODE_LOG_SIZE: u64 = 1;
+pub const CHUNK_LOG_SIZE: usize = 3;
+pub const NODE_LOG_SIZE: usize = 1;
